Convert rooms table and factor out lots of functions
