+++ roomserver/storage/sqlite3/storage.go
@@ -18,14 +18,12 @@ package sqlite3
 import (
 	"context"
 	"database/sql"
-	"encoding/json"
 	"errors"
 	"net/url"
 
 	"github.com/matrix-org/dendrite/internal/sqlutil"
 
 	"github.com/matrix-org/dendrite/internal"
-	"github.com/matrix-org/dendrite/roomserver/api"
 	"github.com/matrix-org/dendrite/roomserver/storage/shared"
 	"github.com/matrix-org/dendrite/roomserver/storage/tables"
 	"github.com/matrix-org/dendrite/roomserver/types"
@@ -110,126 +108,6 @@ func Open(dataSourceName string) (*Database, error) {
 	return &d, nil
 }
 
-// StoreEvent implements input.EventDatabase
-func (d *Database) StoreEvent(
-	ctx context.Context, event gomatrixserverlib.Event,
-	txnAndSessionID *api.TransactionID, authEventNIDs []types.EventNID,
-) (types.RoomNID, types.StateAtEvent, error) {
-	var (
-		roomNID          types.RoomNID
-		eventTypeNID     types.EventTypeNID
-		eventStateKeyNID types.EventStateKeyNID
-		eventNID         types.EventNID
-		stateNID         types.StateSnapshotNID
-		err              error
-	)
-
-	err = internal.WithTransaction(d.db, func(txn *sql.Tx) error {
-		if txnAndSessionID != nil {
-			if err = d.transactions.InsertTransaction(
-				ctx, txn, txnAndSessionID.TransactionID,
-				txnAndSessionID.SessionID, event.Sender(), event.EventID(),
-			); err != nil {
-				return err
-			}
-		}
-
-		// TODO: Here we should aim to have two different code paths for new rooms
-		// vs existing ones.
-
-		// Get the default room version. If the client doesn't supply a room_version
-		// then we will use our configured default to create the room.
-		// https://matrix.org/docs/spec/client_server/r0.6.0#post-matrix-client-r0-createroom
-		// Note that the below logic depends on the m.room.create event being the
-		// first event that is persisted to the database when creating or joining a
-		// room.
-		var roomVersion gomatrixserverlib.RoomVersion
-		if roomVersion, err = extractRoomVersionFromCreateEvent(event); err != nil {
-			return err
-		}
-
-		if roomNID, err = d.assignRoomNID(ctx, txn, event.RoomID(), roomVersion); err != nil {
-			return err
-		}
-
-		if eventTypeNID, err = d.assignEventTypeNID(ctx, txn, event.Type()); err != nil {
-			return err
-		}
-
-		eventStateKey := event.StateKey()
-		// Assigned a numeric ID for the state_key if there is one present.
-		// Otherwise set the numeric ID for the state_key to 0.
-		if eventStateKey != nil {
-			if eventStateKeyNID, err = d.assignStateKeyNID(ctx, txn, *eventStateKey); err != nil {
-				return err
-			}
-		}
-
-		if eventNID, stateNID, err = d.events.InsertEvent(
-			ctx,
-			txn,
-			roomNID,
-			eventTypeNID,
-			eventStateKeyNID,
-			event.EventID(),
-			event.EventReference().EventSHA256,
-			authEventNIDs,
-			event.Depth(),
-		); err != nil {
-			if err == sql.ErrNoRows {
-				// We've already inserted the event so select the numeric event ID
-				eventNID, stateNID, err = d.events.SelectEvent(ctx, txn, event.EventID())
-			}
-			if err != nil {
-				return err
-			}
-		}
-
-		if err = d.eventJSON.InsertEventJSON(ctx, txn, eventNID, event.JSON()); err != nil {
-			return err
-		}
-
-		return nil
-	})
-	if err != nil {
-		return 0, types.StateAtEvent{}, err
-	}
-
-	return roomNID, types.StateAtEvent{
-		BeforeStateSnapshotNID: stateNID,
-		StateEntry: types.StateEntry{
-			StateKeyTuple: types.StateKeyTuple{
-				EventTypeNID:     eventTypeNID,
-				EventStateKeyNID: eventStateKeyNID,
-			},
-			EventNID: eventNID,
-		},
-	}, nil
-}
-
-func extractRoomVersionFromCreateEvent(event gomatrixserverlib.Event) (
-	gomatrixserverlib.RoomVersion, error,
-) {
-	var err error
-	var roomVersion gomatrixserverlib.RoomVersion
-	// Look for m.room.create events.
-	if event.Type() != gomatrixserverlib.MRoomCreate {
-		return gomatrixserverlib.RoomVersion(""), nil
-	}
-	roomVersion = gomatrixserverlib.RoomVersionV1
-	var createContent gomatrixserverlib.CreateContent
-	// The m.room.create event contains an optional "room_version" key in
-	// the event content, so we need to unmarshal that first.
-	if err = json.Unmarshal(event.Content(), &createContent); err != nil {
-		return gomatrixserverlib.RoomVersion(""), err
-	}
-	// A room version was specified in the event content?
-	if createContent.RoomVersion != nil {
-		roomVersion = gomatrixserverlib.RoomVersion(*createContent.RoomVersion)
-	}
-	return roomVersion, err
-}
-
 func (d *Database) assignRoomNID(
 	ctx context.Context, txn *sql.Tx,
 	roomID string, roomVersion gomatrixserverlib.RoomVersion,
@@ -242,22 +120,6 @@ func (d *Database) assignRoomNID(
 		if err == nil {
 			// Now get the numeric ID back out of the database
 			roomNID, err = d.rooms.SelectRoomNID(ctx, txn, roomID)
-		}
-	}
-	return
-}
-
-func (d *Database) assignEventTypeNID(
-	ctx context.Context, txn *sql.Tx, eventType string,
-) (eventTypeNID types.EventTypeNID, err error) {
-	// Check if we already have a numeric ID in the database.
-	eventTypeNID, err = d.eventTypes.SelectEventTypeNID(ctx, txn, eventType)
-	if err == sql.ErrNoRows {
-		// We don't have a numeric ID so insert one into the database.
-		eventTypeNID, err = d.eventTypes.InsertEventTypeNID(ctx, txn, eventType)
-		if err == sql.ErrNoRows {
-			// We raced with another insert so run the select again.
-			eventTypeNID, err = d.eventTypes.SelectEventTypeNID(ctx, txn, eventType)
 		}
 	}
 	return
@@ -277,47 +139,6 @@ func (d *Database) assignStateKeyNID(
 		}
 	}
 	return
-}
-
-// Events implements input.EventDatabase
-func (d *Database) Events(
-	ctx context.Context, eventNIDs []types.EventNID,
-) ([]types.Event, error) {
-	var eventJSONs []tables.EventJSONPair
-	var err error
-	var results []types.Event
-	eventJSONs, err = d.eventJSON.BulkSelectEventJSON(ctx, eventNIDs)
-	if err != nil || len(eventJSONs) == 0 {
-		return nil, nil
-	}
-	err = internal.WithTransaction(d.db, func(txn *sql.Tx) error {
-		results = make([]types.Event, len(eventJSONs))
-		for i, eventJSON := range eventJSONs {
-			var roomNID types.RoomNID
-			var roomVersion gomatrixserverlib.RoomVersion
-			result := &results[i]
-			result.EventNID = eventJSON.EventNID
-			roomNID, err = d.events.SelectRoomNIDForEventNID(ctx, txn, eventJSON.EventNID)
-			if err != nil {
-				return err
-			}
-			roomVersion, err = d.rooms.SelectRoomVersionForRoomNID(ctx, txn, roomNID)
-			if err != nil {
-				return err
-			}
-			result.Event, err = gomatrixserverlib.NewEventFromTrustedJSON(
-				eventJSON.EventJSON, false, roomVersion,
-			)
-			if err != nil {
-				return nil
-			}
-		}
-		return nil
-	})
-	if err != nil {
-		return []types.Event{}, err
-	}
-	return results, nil
 }
 
 // AddState implements input.EventDatabase
@@ -789,21 +610,6 @@ func (d *Database) GetMembershipEventNIDsForRoom(
 		return nil
 	})
 	return
-}
-
-// EventsFromIDs implements query.RoomserverQueryAPIEventDB
-func (d *Database) EventsFromIDs(ctx context.Context, eventIDs []string) ([]types.Event, error) {
-	nidMap, err := d.EventNIDs(ctx, eventIDs)
-	if err != nil {
-		return nil, err
-	}
-
-	var nids []types.EventNID
-	for _, nid := range nidMap {
-		nids = append(nids, nid)
-	}
-
-	return d.Events(ctx, nids)
 }
 
 type transaction struct {
